Refactors `ScanningScreen` to common code and integrates cameraK library
This commit moves `ScanningScreen` to the commonMain source set, allowing for shared code between platforms. It also integrates the `cameraK` and `qr_scanner_plugin` libraries for camera and QR scanning functionalities.
