+++ composeApp/src/androidMain/kotlin/me/androidbox/qrcraft/presentation/CameraPreviewContent.kt
@@ -9,6 +9,7 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.compose.LocalLifecycleOwner
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import me.androidbox.qrcraft.presentation.CameraPreviewViewModel
 
 @Composable
 fun CameraPreviewContent(
